Added several MicrobiomeDB identified results
MicrobiomeDB collected over 1000 samples used for study of diarrheal
disease and the microbiome in children. Using the dataset and the
MicrobiomeDB data and data analysis tool, we found that Prevotella
bacteria is significantly lower in diarrhea patients compared to
control; however, unclassified Proteobacteria and Natronobacillus
bacteria statistically increased in diarrhea patients compared to
control. These are represented in OHMI.